simple_lmk: Replace waitqueue with manual scheduling
Put a simple scheduling loop with a post-while condition to get rid of unnecessary overhead the waitq structure brings specially we only need the thread to wake up and sleep without any timer needed whatsoever. Let's also set the task's state as TASK_IDLE to be excluded from loadavg when it schedules as the thread does nothing aside from wait for the wake up signal.

Signed-off-by: ederekun <sedrickvince@gmail.com>